salbot-master/requirements.txt
discord.py
expiring-dict
aiosqlite
python-dotenv
mcstatus